MJS: Add device-tree overlay for the HTU21D sensor chip
This is the sensor chip used in the MJS measurement station, which we
can conveniently attach to the I²C grove port of the Beagleboard Green
inside he Lorank.

This configuration adds a device-tree overlay that tells the kernel
about this sensor and its address, which loads a driver and makes it
available at /sys/bus/iio (or /sys/bus/i2c/drivers/htu21).