configure_gateway: Be smarter about password changes
Previously, it would just always fetch a password from the password
store (causing a gpg keyphrase popup) and then always setting the
password.

Now, it does this when the password is still at default. If so, and the
password is not available, the password is disabled instead (to not
leave the system vulnerabl). If the password then is available on
a later run, the disabled password is replaced by the proper password
again.