nuttssh: Use StrictHostKeyChecking=accept-new instead of no
The idea was to use "no" so a changed host key is accepted without
question. However, it turns out that when the host key changed, there
are some restrictions. One restriction is that port forwarding is
disabled, which breaks the nuttssh connection entirely.

Because of this, there is no point in using "no", so change to
"accept-new". This still allows automatically adding a new key on
deployment (or on key rotation if asyncssh implements
https://github.com/ronf/asyncssh/issues/667).

This means that when changing the server, the keys should be preserved,
which is ok. It also removes the risk of a man-in-the-middle attack
against the nuttssh client (though it has limited impact).